[CONFIGURATION] * NEW: Print a summary of warning messages after each component   compilation.
[INTERPRETER]
* BUG: StrPtr() with a length argument now works correctly.
* NEW: Support for callbacks: when converting a function to a Pointer, an 
  internal closure is created with the libffi library. The pointer then can 
  be sent to any extern function as function pointer argument.
* BUG: Opening memory stream should work in all cases now.


git-svn-id: svn://localhost/gambas/trunk@3038 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ gb.db.postgresql/src/main.c
@@ -175,7 +175,7 @@ static char *get_quoted_table(const char *table)
 	else
 	{
 		res = GB.TempString(NULL, len + 2);
-		sprintf(res, "%.*s.\"%s\"", point - table, table, point + 1);
+		sprintf(res, "%.*s.\"%s\"", (int)(point - table), table, point + 1);
 	}
 	
 	return res;
